third_party/blink/web_tests/platform/mac-mac10.10/external/wpt/webxr/xrView_eyes.https-expected.txt

This is a testharness.js-based test.
PASS XRView.eye is correct for immersive sessions
PASS XRView.eye is correct for non-immersive sessions
Harness: the test ran to completion.
